convert to typescript, add support for sponsor pools
