--- elasticdao-config.json
@@ -1,8 +1,0 @@
-{
-  "contestId": 2,
-  "sponsorName": "elasticdao",
-  "mainPool": 20,
-  "gasPool": 2,
-  "awardCoin": "ETH",
-  "awardCoinInUSD": 1800
-}
