--- elasticdao-judged.json
@@ -1,300 +1,0 @@
-[ 
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "H-01",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "H-02",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "H-03",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "H-04",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "H-05",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "L-01",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "L-02",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "L-03",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "L-04",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "L-09",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "M-01",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "cmichel",
-    "finding": "M-05",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "H-05",
-    "risk": "3"
-  },{ 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "L-05",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "L-06",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "L-08",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "L-12",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "janbro",
-    "finding": "H-05",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "janbro",
-    "finding": "L-07",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "janbro",
-    "finding": "L-10",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "ncitron",
-    "finding": "H-05",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "H-08",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "H-05",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "H-07",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "L-02",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "L-08",
-    "risk": "1"
-  },{ 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "M-02",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "N-02",
-    "risk": "0"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "H-06",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "H-07",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "L-08",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "L-11",
-    "risk": "1"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "M-03",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "M-04",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "M-06",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "M-07",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  { 
-    "contest": "2",
-    "handle": "s1m0",
-    "finding": "H-07",
-    "risk": "3"
-  },
-  { 
-    "contest": "2",
-    "handle": "s1m0",
-    "finding": "M-01",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "s1m0",
-    "finding": "M-02",
-    "risk": "2"
-  },
-  { 
-    "contest": "2",
-    "handle": "gpersoon",
-    "finding": "G-06",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "G-01",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "G-02",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "G-03",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "G-04",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pauliax",
-    "finding": "G-05",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "G-01",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "pocotiempo",
-    "finding": "G-05",
-    "risk": "g"
-  },
-  { 
-    "contest": "2",
-    "handle": "s1m0",
-    "finding": "G-07",
-    "risk": "g"
-  }
-]
