--- slingshot-judged.json
@@ -1,254 +1,0 @@
-[
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "H-01",
-    "risk": "3"
-  },
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "L-02",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "M-01",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "M-02",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "M-05",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "pocotiempo",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "G-06",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-04",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-06",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-07",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-08",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-09",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "L-10",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "M-04",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "pauliax",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "G-01",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "G-02",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "G-03",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "G-04",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "L-12",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "onewayfunction",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  {
-    "contest": "1",
-    "handle": "nedodn",
-    "finding": "G-03",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "nedodn",
-    "finding": "M-03",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "ncitron",
-    "finding": "L-04",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "ncitron",
-    "finding": "L-05",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "ncitron",
-    "finding": "M-02",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "janbro",
-    "finding": "G-02",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "janbro",
-    "finding": "M-01",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "janbro",
-    "finding": "M-02",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "L-02",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "L-03",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "L-06",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "L-11",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "M-04",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "N-01",
-    "risk": "0"
-  },
-  {
-    "contest": "1",
-    "handle": "gpersoon",
-    "finding": "N-03",
-    "risk": "0"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "G-01",
-    "risk": "g"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "L-01",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "L-02",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "L-03",
-    "risk": "1"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "M-01",
-    "risk": "2"
-  },
-  {
-    "contest": "1",
-    "handle": "cmichel",
-    "finding": "N-01",
-    "risk": "0"
-  }
-]
